Exception handling.

+++ src/main/java/com/tatyanajacques/dscommerce/services/ProductService.java
@@ -1,16 +1,19 @@
 package com.tatyanajacques.dscommerce.services;
 
+
 import com.tatyanajacques.dscommerce.dto.ProductDTO;
 import com.tatyanajacques.dscommerce.entities.Product;
 import com.tatyanajacques.dscommerce.repositories.ProductRepository;
+import com.tatyanajacques.dscommerce.services.exceptions.DatabaseException;
+import com.tatyanajacques.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-import java.util.Optional;
 
 @Service
 public class ProductService {
@@ -20,8 +23,10 @@ public class ProductService {
 
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id){
-        Product product = repository.findById(id).get();
-        return new ProductDTO(product);
+            Product product = repository.findById(id).orElseThrow(
+                    ()-> new ResourceNotFoundException("Recurso não encontrado."));
+            return new ProductDTO(product);
+
     }
 
     @Transactional(readOnly = true)
@@ -40,15 +45,32 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto){
-        Product entity = repository.getReferenceById(id);
-        copyDTOToEntity(dto, entity);
-        entity = repository.save(entity);
-        return new ProductDTO(entity);
+        try{
+            Product entity = repository.getReferenceById(id);
+            copyDTOToEntity(dto, entity);
+            entity = repository.save(entity);
+            return new ProductDTO(entity);
+        }
+        catch(EntityNotFoundException e ){
+            throw new ResourceNotFoundException("Recurso não encontrado.");
+        }
+
+
+
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id){
-        repository.deleteById(id);
+        if(!repository.existsById(id)){
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+       try{
+            repository.deleteById(id);
+        }
+        catch(DataIntegrityViolationException e){
+            throw new DatabaseException("Falha de integridade referencial");
+       }
+
     }
 
     private void copyDTOToEntity(ProductDTO dto, Product entity) {
